What makes a succeesful TED
For a Bachelor Project that we have by end of 5th semester in Germany, Me and my group decided to take a Topic, successful TED Talks. Hence, We desinged a study to find out factors that might be influced to TED talk's scueess.

A standard of successful TED talk is how many views TED veidoe has got in Youtube. And each factors are analyzied by ourselves in a subjective and objective way according to thier character. At the end we came out with 19 different factors in 8 categories.

We decided to apply Pearson R, ANOVA and linear regression according to if the factor is Ordinal or Nominal.